pyproject.toml
[tool.poetry]
name = "money-scheduler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "money_scheduler"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
django-bootstrap5 = "^23.4"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
